Extract run-tests task. Save summary as metadata
Decouples running the tests from what to do on test errors or failures.
`test` is now a task composed of `run-tests` which runs the tests and
saves the summary as metadata, and an inline task that throws an
exception when it finds errors or failures in that summary.

Having the test summary as fileset metadata will allow for further
processing in downstream tasks.